♻️ [Refactor] : History 페이지의 연혁 카드 전체 section 컴포넌트화
+ HistorySwiper.jsx file 을 생성하여, History 페이지 내에서 개발되었던 Section 전체를 컴포넌트화 하였습니다.

Related to: #6
